move useStore out of core
https://github.com/transloadit/uppy/pull/5443#discussion_r1862342884
(cherry picked from commit 00070c870f76876aca92e97f410f8b52085191bc)
